services/paddleocr/pyproject.toml
[project]
name = "paddleocr-service"
version = "0.1.0"
description = "paddleocr recognition service with observabilitye"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.3",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-exporter-otlp>=1.38.0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "opentelemetry-sdk>=1.38.0",
    "paddleocr>=3.3.2",
    "paddlepaddle>=3.2.2",
    "pillow>=12.0.0",
    "prometheus-client>=0.23.1",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.20",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.4.0",
    "ruff>=0.14.6",
]

[tool.ruff]
line-length = 100
target-version = "py313"
extend-exclude = [
    ".venv",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
# structlog processors
"src/logger.py" = ["ARG001"]
# fastapi lifespan app
"src/main.py" = ["ARG001"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
